requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
supabase
slowapi==0.1.9
accelerate
transformers